\name{bcv-package}
\alias{bcv-package}
\alias{bcv}
\docType{package}
\title{
    Bi-Cross Validation
}
\description{
    Implements cross validation for choosing the rank of an SVD approximation
    to a data matrix.
}
\details{
\tabular{ll}{
Package: \tab bcv\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-09-18\cr
License: \tab BSD3\cr
}

Basic usage is to call the "cv.svd" function.

}
\author{
Patrick O. Perry

Maintainer: Patrick O. Perry <patperry@stanford.edu> 
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
