% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.kmeans.gabriel.R
\name{cv.kmeans.gabriel}
\alias{cv.kmeans.gabriel}
\title{Choosing the number of clusters of a matrix using Gabriel CV method.}
\usage{
cv.kmeans.gabriel(x, krow = 5, kcol = 2, maxcenters = 10,
  classify.method = c("nearest", "lda-equal", "lda-proportions"))
}
\arguments{
\item{x}{The matrix to find the number of clusters.}

\item{krow}{The number of row folds.}

\item{kcol}{The number of column folds.}

\item{maxcenters}{The upper bound of searching range, i.e k searched in 1:maxcenters}

\item{classify.method}{The classifier used inside the algorithm,
with possible choice "nearest", "lda-equal" and "lda-proportions".}
}
\value{
\item{msep}{The mean square error of prediction (MSEP); this is a
 matrix whose columns contain the mean square errors in the predictions of the holdout
 sets for k = 1, ..., maxcenters across the different replicates.}
\item{centers}{The number of clusters chosen by the algorithm.}
}
\description{
Perform Gabriel-style cross-validation for determining the
number of clusters of a matrix.
}
\details{
Through gabriel cross-validation, the data matrix has been decomposed
into four blocks:
\deqn{\mathbf{X} = \begin{bmatrix} X_{train} & Y_{train} \\ X_{test}  & Y_{test} \end{bmatrix},}
where \eqn{X} and \eqn{Y} are formed by CV on the column. Note here \eqn{Y} is not the
real responses since the problem is a unsupervised learning problem. Instead, \eqn{Y}
is the leaf-out part of each CV (on the column ) replicate while \eqn{X} is the leaf-in part.

For each CV replicate, the algorithm applies kmeans with parameter k = 1, ..., maxcenters on
\eqn{Y_{train}}, which yields k clusters' assigment; builds a classifier using the training
data (i.e. \eqn{X_{train}} and \eqn{Y_{train}}) where \eqn{Y_{train}} has been replaced by
the cluster assigments; predicts the cluster assigments of \eqn{X_{test}} using the
classifier just trained; and calculates the prediction error by measuring the difference
between \eqn{Y_{test}} and the predicted cluster assigments (each cluster
assigment is replaced by the correspoding cluster center). The overall prediction
error is average across all CV replicates and the algorithm picks the k
that minimizes such error.
}
\examples{
## generate a 40x2 matrix with two clusters centered at (4,4) and (0,0)
cluster1 <- matrix(c(4,4),nrow = 20, ncol = 2)+ matrix(rnorm(40),nrow = 20, ncol = 2)
cluster2 <- matrix(c(0,0),nrow = 20, ncol = 2)+ matrix(rnorm(40),nrow = 20, ncol = 2)
x = rbind(cluster1,cluster2 )

## cluster the matrix with gabriel CV method with 5 by 2 cross-validation
cv.kmeans.gabriel(x, 5, 2, maxcenters=10, classify.method="nearest")

}
\author{
Wei Fu, Patrick O. Perry
}

